Ajout du systeme de tir avec tentative de pathfinding

+++ Dual-Punk/Assets/Scripts/Entities/Player/PlayerState.cs
@@ -35,7 +35,7 @@ public class PlayerState : MonoBehaviour
         Destroy(gameObject);
     }
 
-    public void Damage(int amount){
+    public void Damage(float amount){
         if (amount < 0){
             throw new ArgumentOutOfRangeException("Cannot have Negative damage.");
         }
